refactor: cleaning of using package/global defaults
