+++ cachier/__init__.py
@@ -1,10 +1,10 @@
-from .core import (
-    cachier,
+from .config import (
     set_default_params,
     get_default_params,
     enable_caching,
     disable_caching,
 )
+from .core import cachier
 
 from ._version import *  # noqa: F403
