+++ cachier/cores/base.py
@@ -9,6 +9,8 @@
 import abc  # for the _BaseCore abstract base class
 import inspect
 
+from ..config import _Type_HashFunc
+
 
 class RecalculationNeeded(Exception):
     pass
@@ -17,9 +19,9 @@ class RecalculationNeeded(Exception):
 class _BaseCore:
     __metaclass__ = abc.ABCMeta
 
-    def __init__(self, hash_func, default_params):
-        self.default_params = default_params
+    def __init__(self, hash_func: _Type_HashFunc, wait_for_calc_timeout: int):
         self.hash_func = hash_func
+        self.wait_for_calc_timeout = wait_for_calc_timeout
 
     def set_func(self, func):
         """Sets the function this core will use.
@@ -37,10 +39,7 @@ class _BaseCore:
 
     def get_key(self, args, kwds):
         """Returns a unique key based on the arguments provided."""
-        if self.hash_func is not None:
-            return self.hash_func(args, kwds)
-        else:
-            return self.default_params["hash_func"](args, kwds)
+        return self.hash_func(args, kwds)
 
     def get_entry(self, args, kwds):
         """Returns the result mapped to the given arguments in this core's
@@ -56,10 +55,7 @@ class _BaseCore:
 
     def check_calc_timeout(self, time_spent):
         """Raise an exception if a recalculation is needed."""
-        if self.wait_for_calc_timeout is not None:
-            calc_timeout = self.wait_for_calc_timeout
-        else:
-            calc_timeout = self.default_params["wait_for_calc_timeout"]
+        calc_timeout = self.wait_for_calc_timeout
         if calc_timeout > 0 and (time_spent >= calc_timeout):
             raise RecalculationNeeded()
